+++ contracts/Exchange.sol
@@ -247,11 +247,17 @@ contract Exchange is
         // EX_BTNE: base token does not exists
         require(IMarketRegistry(_marketRegistry).hasPool(baseToken), "EX_BTNE");
 
-        // if updating TWAP fails, this call will be reverted and thus using try-catch
-        try IBaseToken(baseToken).cacheTwap(IClearingHouseConfig(_clearingHouseConfig).getTwapInterval()) {} catch {}
-        uint256 markTwap;
+        // The purpose of caching index twap here is to save the gas consumption of calculating mark price,
+        // if updating TWAP fails, this call will be reverted and thus using try-catch.
+        // NOTE: the cached index twap is used for AccountBalance.MarkPrice calculation,
+        // not for funding rate calculation.
+        (, uint32 premiumInterval) = IClearingHouseConfig(_clearingHouseConfig).getMarkPriceConfig();
+        try IBaseToken(baseToken).cacheTwap(premiumInterval) {} catch {}
+
+        // markTwap here is market twap, and used for funding rate calculation.
+        uint256 marketTwap;
         uint256 indexTwap;
-        (fundingGrowthGlobal, markTwap, indexTwap) = _getFundingGrowthGlobalAndTwaps(baseToken);
+        (fundingGrowthGlobal, marketTwap, indexTwap) = _getFundingGrowthGlobalAndTwaps(baseToken);
 
         fundingPayment = _updateFundingGrowth(
             trader,
@@ -275,7 +281,7 @@ contract Exchange is
                 lastFundingGrowthGlobal.twPremiumDivBySqrtPriceX96
             ) = (timestamp, fundingGrowthGlobal.twPremiumX96, fundingGrowthGlobal.twPremiumDivBySqrtPriceX96);
 
-            emit FundingUpdated(baseToken, markTwap, indexTwap);
+            emit FundingUpdated(baseToken, marketTwap, indexTwap);
 
             // update tick for price limit checks
             _lastUpdatedTickMap[baseToken] = _getTick(baseToken);
@@ -346,11 +352,22 @@ contract Exchange is
 
     /// @inheritdoc IExchange
     function isOverPriceSpread(address baseToken) external view override returns (bool) {
-        uint256 markPrice = getSqrtMarkTwapX96(baseToken, 0).formatSqrtPriceX96ToPriceX96().formatX96ToX10_18();
+        uint256 marketPrice = _getSqrtMarketTwapX96(baseToken, 0).formatSqrtPriceX96ToPriceX96().formatX96ToX10_18();
         uint256 indexTwap =
             IIndexPrice(baseToken).getIndexPrice(IClearingHouseConfig(_clearingHouseConfig).getTwapInterval());
-        uint256 spread = markPrice > indexTwap ? markPrice.sub(indexTwap) : indexTwap.sub(markPrice);
+        uint256 spread = marketPrice > indexTwap ? marketPrice.sub(indexTwap) : indexTwap.sub(marketPrice);
         return spread > PerpMath.mulRatio(indexTwap, _MAX_PRICE_SPREAD_RATIO);
+    }
+
+    /// @inheritdoc IExchange
+    // Deprecated function, will be removed in the next release
+    function getSqrtMarkTwapX96(address baseToken, uint32 twapInterval) external view override returns (uint160) {
+        return _getSqrtMarketTwapX96(baseToken, twapInterval);
+    }
+
+    /// @inheritdoc IExchange
+    function getSqrtMarketTwapX96(address baseToken, uint32 twapInterval) external view override returns (uint160) {
+        return _getSqrtMarketTwapX96(baseToken, twapInterval);
     }
 
     //
@@ -371,11 +388,6 @@ contract Exchange is
                 fundingGrowthGlobal,
                 liquidityCoefficientInFundingPayment
             );
-    }
-
-    /// @inheritdoc IExchange
-    function getSqrtMarkTwapX96(address baseToken, uint32 twapInterval) public view override returns (uint160) {
-        return UniswapV3Broker.getSqrtMarkTwapX96(IMarketRegistry(_marketRegistry).getPool(baseToken), twapInterval);
     }
 
     //
@@ -570,6 +582,10 @@ contract Exchange is
     // INTERNAL VIEW
     //
 
+    function _getSqrtMarketTwapX96(address baseToken, uint32 twapInterval) internal view returns (uint160) {
+        return UniswapV3Broker.getSqrtMarketTwapX96(IMarketRegistry(_marketRegistry).getPool(baseToken), twapInterval);
+    }
+
     function _isOverPriceLimit(address baseToken) internal view returns (bool) {
         int24 tick = _getTick(baseToken);
         return _isOverPriceLimitWithTick(baseToken, tick);
@@ -591,14 +607,14 @@ contract Exchange is
 
     /// @dev this function calculates the up-to-date globalFundingGrowth and twaps and pass them out
     /// @return fundingGrowthGlobal the up-to-date globalFundingGrowth
-    /// @return markTwap only for settleFunding()
+    /// @return marketTwap only for settleFunding()
     /// @return indexTwap only for settleFunding()
     function _getFundingGrowthGlobalAndTwaps(address baseToken)
         internal
         view
         returns (
             Funding.Growth memory fundingGrowthGlobal,
-            uint256 markTwap,
+            uint256 marketTwap,
             uint256 indexTwap
         )
     {
@@ -615,9 +631,9 @@ contract Exchange is
             twapInterval = twapInterval > deltaTimestamp ? deltaTimestamp : twapInterval;
         }
 
-        uint256 markTwapX96;
+        uint256 marketTwapX96;
         if (marketOpen) {
-            markTwapX96 = getSqrtMarkTwapX96(baseToken, twapInterval).formatSqrtPriceX96ToPriceX96();
+            marketTwapX96 = _getSqrtMarketTwapX96(baseToken, twapInterval).formatSqrtPriceX96ToPriceX96();
             indexTwap = IIndexPrice(baseToken).getIndexPrice(twapInterval);
         } else {
             // if a market is paused/closed, we use the last known index price which is getPausedIndexPrice
@@ -627,12 +643,12 @@ contract Exchange is
 
             // timestamp is pausedTime when the market is not open
             uint32 secondsAgo = _blockTimestamp().sub(timestamp).toUint32();
-            markTwapX96 = UniswapV3Broker
-                .getSqrtMarkTwapX96From(IMarketRegistry(_marketRegistry).getPool(baseToken), secondsAgo, twapInterval)
+            marketTwapX96 = UniswapV3Broker
+                .getSqrtMarketTwapX96From(IMarketRegistry(_marketRegistry).getPool(baseToken), secondsAgo, twapInterval)
                 .formatSqrtPriceX96ToPriceX96();
             indexTwap = IBaseToken(baseToken).getPausedIndexPrice();
         }
-        markTwap = markTwapX96.formatX96ToX10_18();
+        marketTwap = marketTwapX96.formatX96ToX10_18();
 
         uint256 lastSettledTimestamp = _lastSettledTimestampMap[baseToken];
         Funding.Growth storage lastFundingGrowthGlobal = _globalFundingGrowthX96Map[baseToken];
@@ -640,9 +656,9 @@ contract Exchange is
             // if this is the latest updated timestamp, values in _globalFundingGrowthX96Map are up-to-date already
             fundingGrowthGlobal = lastFundingGrowthGlobal;
         } else {
-            // deltaTwPremium = (markTwap - indexTwap) * (now - lastSettledTimestamp)
+            // deltaTwPremium = (marketTwap - indexTwap) * (now - lastSettledTimestamp)
             int256 deltaTwPremiumX96 =
-                _getDeltaTwapX96(markTwapX96, indexTwap.formatX10_18ToX96()).mul(
+                _getDeltaTwapX96(marketTwapX96, indexTwap.formatX10_18ToX96()).mul(
                     timestamp.sub(lastSettledTimestamp).toInt256()
                 );
             fundingGrowthGlobal.twPremiumX96 = lastFundingGrowthGlobal.twPremiumX96.add(deltaTwPremiumX96);
@@ -650,13 +666,13 @@ contract Exchange is
             // overflow inspection:
             // assuming premium = 1 billion (1e9), time diff = 1 year (3600 * 24 * 365)
             // log(1e9 * 2^96 * (3600 * 24 * 365) * 2^96) / log(2) = 246.8078491997 < 255
-            // twPremiumDivBySqrtPrice += deltaTwPremium / getSqrtMarkTwap(baseToken)
+            // twPremiumDivBySqrtPrice += deltaTwPremium / getSqrtMarketTwap(baseToken)
             fundingGrowthGlobal.twPremiumDivBySqrtPriceX96 = lastFundingGrowthGlobal.twPremiumDivBySqrtPriceX96.add(
-                PerpMath.mulDiv(deltaTwPremiumX96, PerpFixedPoint96._IQ96, getSqrtMarkTwapX96(baseToken, 0))
+                PerpMath.mulDiv(deltaTwPremiumX96, PerpFixedPoint96._IQ96, _getSqrtMarketTwapX96(baseToken, 0))
             );
         }
 
-        return (fundingGrowthGlobal, markTwap, indexTwap);
+        return (fundingGrowthGlobal, marketTwap, indexTwap);
     }
 
     /// @dev get a price limit for replaySwap s.t. it can stop when reaching the limit to save gas
@@ -675,15 +691,15 @@ contract Exchange is
         return TickMath.getSqrtRatioAtTick(tickBoundary);
     }
 
-    function _getDeltaTwapX96(uint256 markTwapX96, uint256 indexTwapX96) internal view returns (int256 deltaTwapX96) {
+    function _getDeltaTwapX96(uint256 marketTwapX96, uint256 indexTwapX96) internal view returns (int256 deltaTwapX96) {
         uint24 maxFundingRate = IClearingHouseConfig(_clearingHouseConfig).getMaxFundingRate();
         uint256 maxDeltaTwapX96 = indexTwapX96.mulRatio(maxFundingRate);
         uint256 absDeltaTwapX96;
-        if (markTwapX96 > indexTwapX96) {
-            absDeltaTwapX96 = markTwapX96.sub(indexTwapX96);
+        if (marketTwapX96 > indexTwapX96) {
+            absDeltaTwapX96 = marketTwapX96.sub(indexTwapX96);
             deltaTwapX96 = absDeltaTwapX96 > maxDeltaTwapX96 ? maxDeltaTwapX96.toInt256() : absDeltaTwapX96.toInt256();
         } else {
-            absDeltaTwapX96 = indexTwapX96.sub(markTwapX96);
+            absDeltaTwapX96 = indexTwapX96.sub(marketTwapX96);
             deltaTwapX96 = absDeltaTwapX96 > maxDeltaTwapX96 ? maxDeltaTwapX96.neg256() : absDeltaTwapX96.neg256();
         }
     }